refactor: aplicar arquitectura hexagonal a datos personales
