+++ src/stores/personal.ts
@@ -1,28 +1,33 @@
 import { defineStore } from 'pinia'
 import { computed } from 'vue'
-import type { PersonalData } from '../interfaces'
-import personalData from '../data/personal.json'
+import type { PersonalData } from '../domain/personal/Personal'
+import { GetPersonalUseCase } from '../application/personal/GetPersonalUseCase'
+import { StaticPersonalRepository } from '../infrastructure/personal/StaticPersonalRepository'
 
 export const usePersonalStore = defineStore('personal', () => {
-  // State
-  const personal: PersonalData = personalData
+  // Instanciamos el repositorio y el caso de uso (capa de infraestructura -> aplicación)
+  const repository = new StaticPersonalRepository()
+  const getPersonalUseCase = new GetPersonalUseCase(repository)
 
-  // Getters
+  // Estado: datos personales obtenidos a través del caso de uso
+  const personal: PersonalData = getPersonalUseCase.execute()
+
+  // Getters: exponen la información al exterior
   const getPersonal = computed(() => personal)
-  
+
   const getProfile = computed(() => personal.profile)
-  
+
   const getContact = computed(() => personal.contact)
-  
+
   const getAbout = computed(() => personal.about)
-  
+
   const getAchievements = computed(() => personal.achievements)
 
   const getGoals = computed(() => personal.about.goals)
-  
+
   const getInterests = computed(() => personal.about.interests)
 
-  // Actions
+  // Actions: métodos que procesan o filtran la información
   const getContactInfo = () => {
     return {
       email: personal.contact.email,
@@ -42,9 +47,9 @@ export const usePersonalStore = defineStore('personal', () => {
   }
 
   return {
-    // State
+    // Estado
     personal,
-    
+
     // Getters
     getPersonal,
     getProfile,
@@ -53,8 +58,8 @@ export const usePersonalStore = defineStore('personal', () => {
     getAchievements,
     getGoals,
     getInterests,
-    
-    // Actions
+
+    // Acciones
     getContactInfo,
     getProfileDescription,
     getGoalsByPeriod
